Refactored 'codegen' dir to move OperatorTranslator classes into 'codegen/operator' #690
