+++ src/codegen/translator_factory.cpp
@@ -17,14 +17,14 @@
 #include "include/codegen/expression/comparison_translator.h"
 #include "include/codegen/expression/conjunction_translator.h"
 #include "include/codegen/expression/constant_translator.h"
-#include "codegen/delete_translator.h"
-#include "codegen/global_group_by_translator.h"
-#include "codegen/hash_group_by_translator.h"
+#include "include/codegen/operator/delete_translator.h"
+#include "include/codegen/operator/global_group_by_translator.h"
+#include "include/codegen/operator/hash_group_by_translator.h"
 #include "codegen/hash_join_translator.h"
 #include "include/codegen/expression/negation_translator.h"
-#include "codegen/order_by_translator.h"
-#include "codegen/projection_translator.h"
-#include "codegen/table_scan_translator.h"
+#include "include/codegen/operator/order_by_translator.h"
+#include "include/codegen/operator/projection_translator.h"
+#include "include/codegen/operator/table_scan_translator.h"
 #include "include/codegen/expression/tuple_value_translator.h"
 #include "expression/case_expression.h"
 #include "expression/comparison_expression.h"
