+++ src/include/codegen/operator/order_by_translator.h
@@ -4,7 +4,7 @@
 //
 // order_by_translator.h
 //
-// Identification: src/include/codegen/order_by_translator.h
+// Identification: src/include/codegen/operator/order_by_translator.h
 //
 // Copyright (c) 2015-17, Carnegie Mellon University Database Group
 //
