+++ src/include/codegen/operator/table_scan_translator.h
@@ -4,7 +4,7 @@
 //
 // table_scan_translator.h
 //
-// Identification: src/include/codegen/table_scan_translator.h
+// Identification: src/include/codegen/operator/table_scan_translator.h
 //
 // Copyright (c) 2015-17, Carnegie Mellon University Database Group
 //
